最新に反映

+++ gt_surveillance.py
@@ -5,11 +5,22 @@ Created on Thu Nov 05 16:44:30 2015
 @brief:  download TCGA bam files
 @author: okada
 
+$Id: gt_surveillance.py 85 2015-12-11 08:04:59Z aokada $
+$Rev: 85 $
+
 # before run
+@code
 export DRMAA_LIBRARY_PATH=/geadmin/N1GE/lib/lx-amd64/libdrmaa.so.1.0
 
 # run
-gt_surveillance.py {path to working dir} {key file} {manifest file} --config_file {option: config file}
+@code
+gt_surveillance.py {path to working dir} {auth key} {manifest file} --config_file {option: config file}
+
+* use --ignore_list option, skip download/
+create this list, for example this command,
+
+ls -l data/ | egrep -v .gto$ | egrep -v .partial$ | cut -f 13 -d ' ' > complete.txt
+@endcode
 """
 
 from multiprocessing import Process
@@ -22,6 +33,7 @@ import drmaa
 import ConfigParser
 import argparse
 import subprocess
+import re
 
 cmd_format = """
 #!/bin/bash
@@ -131,9 +143,9 @@ def qsub_process(name, output_dir, key, manifest, config):
     return return_val
     
 def xml_splitter(name, output_dir, manifest, config):
-
+    prefix = os.path.splitext(os.path.basename(manifest))[0]
     cmd = splitter_format.format(xmlsplitter = config.get('TOOLS', 'xmlsplitter'), \
-                        output_prefix = output_dir + "/manifests/manifest", manifest = manifest)
+                        output_prefix = output_dir + "/manifests/" + prefix, manifest = manifest)
                         
     script_path = output_dir + "/scripts/" + name + ".sh"    
 
@@ -145,23 +157,37 @@ def xml_splitter(name, output_dir, manifest, config):
     p = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     stdout_data, stderr_data = p.communicate()
 
+    li = []
     if p.returncode == 0:
         print "xml_splitter finish: %s" % (stdout_data)
+        li = glob.glob(output_dir + "/manifests/" + prefix + "*")
     else:
         print "xml_splitter error!: %s" % (stderr_data)
 
-    return p.returncode
+    return li
+
+def isbam_downloaded(manifest, bam_dir):
+    f = open(manifest)
+    text = f.read()
+    f.close()
+    analysis = re.split('<analysis_id>|</analysis_id>', text)[1]
+    name = re.split('<filename>|</filename>', text)[1]
+    bam = "%s/%s/%s" % (bam_dir, analysis, name)
+ 
+    return os.path.exists(bam)
     
 def main():
-   
-    # get args
-    parser = argparse.ArgumentParser(prog = "check_bam")
+    name = os.path.splitext(os.path.basename(sys.argv[0]))[0]
     
-    parser.add_argument("--version", action = "version", version = "check_bam-1.0.0")
+    # get args
+    parser = argparse.ArgumentParser(prog = name)
+    
+    parser.add_argument("--version", action = "version", version = name + "-1.0.0")
     parser.add_argument('output_dir', help = "output root directory", type = str)
-    parser.add_argument('key', help = "summary file download from TCGA", type = str)
+    parser.add_argument('key', help = "key file download from TCGA", type = str)
     parser.add_argument('manifest', help = "manifest file downloaded from TCGA", type = str)
     parser.add_argument("--config_file", help = "config file", type = str, default = "")
+    parser.add_argument("--ignore_list", help = "ignore analysis_id list file", type = str, default = "")
     args = parser.parse_args()
     
     output_dir = os.path.abspath(args.output_dir)
@@ -198,19 +224,46 @@ def main():
     # split manifest
     splitter_name = "xml_splitter{0:0>4d}{1:0>2d}{2:0>2d}_{3:0>2d}{4:0>2d}{5:0>2d}".format(
                 now.year, now.month, now.day, now.hour, now.minute, now.second)
-    result = xml_splitter(splitter_name, output_dir, manifest, config)
-    if result != 0:
-        return
+    manifests_org = xml_splitter(splitter_name, output_dir, manifest, config)
     
+    ignore = ""
+    if os.path.exists(args.ignore_list) == True:
+        f = open(args.ignore_list)
+        ignore = f.read()
+        f.close()
+
+    manifests = []
+    for man in manifests_org:
+        f = open(man)
+        data = f.read()
+        f.close()
+
+        man_split = re.split('<analysis_id>|</analysis_id>', data)
+        if len(man_split) < 3:
+            continue
+
+        analysis = man_split[1]
+
+        if ignore.find(analysis) >= 0:
+            write_log(log_path, "a", "Skip this manifest because of ignore list. %s, %s" % (man, analysis), True, True)
+        elif isbam_downloaded(man, output_dir + "/data") == True:
+            write_log(log_path, "a", "Skip this manifest because of already downloaded. %s, %s" % (man, analysis), True, True)
+        else:
+            manifests.append(man)
+
+    manifests.sort()
+    
+    write_log(log_path, "a", "Files = %d." % (len(manifests)), True, True)
+
     process_list = []
-    manifests = glob.glob(output_dir + "/manifests/*")
     if len(manifests) < 1:
         print ("no manifests.")
         
     elif len(manifests) == 1:
         now = datetime.datetime.now()
-        name = "{name}_manifest{index:0>5d}_{year:0>4d}{month:0>2d}{day:0>2d}_{hour:0>2d}{minute:0>2d}{second:0>2d}".format(
-                name = "gt_surveilance", index = 1,
+        man_name = os.path.splitext(os.path.basename(manifest))[0]
+        name = "{name}_{man_name}_{year:0>4d}{month:0>2d}{day:0>2d}_{hour:0>2d}{minute:0>2d}{second:0>2d}".format(
+                name = "gt_surveilance", man_name = man_name,
                 year = now.year, month = now.month, day = now.day, 
                 hour = now.hour, minute = now.minute, second = now.second)
         qsub_process(name, output_dir, key, manifests[0], config)
@@ -237,8 +290,9 @@ def main():
                     
                 manifest = manifests[j]
                 now = datetime.datetime.now()
-                name = "{name}_manifest{index:0>5d}_{year:0>4d}{month:0>2d}{day:0>2d}_{hour:0>2d}{minute:0>2d}{second:0>2d}".format(
-                        name = "gt_surveilance", index = j+1,
+                man_name = os.path.splitext(os.path.basename(manifest))[0]
+                name = "{name}_{man_name}_{year:0>4d}{month:0>2d}{day:0>2d}_{hour:0>2d}{minute:0>2d}{second:0>2d}".format(
+                        name = "gt_surveilance", man_name = man_name,
                         year = now.year, month = now.month, day = now.day, 
                         hour = now.hour, minute = now.minute, second = now.second)
                 process = Process(target=qsub_process, name=name, args=(name, output_dir, key, manifest, config))
@@ -256,14 +310,24 @@ def main():
     for process in process_list:
         process.join()
 
-        f_plog = open(output_dir + '/log/' + process.name + ".log")
+        plog_file = output_dir + '/log/' + process.name + ".log"
+        if os.path.exists(plog_file) == False:
+            continue
+        
+        f_plog = open(plog_file)
         plog = f_plog.read()
         f_plog.close()
         os.remove(output_dir + '/log/' + process.name + ".log")
 
         write_log(log_path, "a", plog, False, False)
 
-    write_log(log_path, "a", "End main process.", True, True)
+    failure = 0
+    for manifest in manifests:
+        if isbam_downloaded(manifest, output_dir + "/data") == False:
+            write_log(log_path, "a", "Failed file to download = %s" % manifest, True, True)
+            failure += 1
+            
+    write_log(log_path, "a", "End main process. Failed file to download = %d" % failure, True, True)
 
 if __name__ == "__main__":
     main()
